chore: make everything work good
